Move Payment Methods management from Invoicing tab to Settings page
Extract PaymentMethodsManager into a shared component and render it
on the Settings page, replacing the Phase 2 placeholder. Remove the
payment methods section from InvoicingTab so the Manage page focuses
on invoices/contacts while Settings consolidates account-level config.

Update InvoicingTab tests to reflect the removed payment methods UI.

Closes #87
